python-service/requirements.txt
Flask
pytesseract
Pillow
PyMuPDF
pdfplumber
werkzeug
